updated bookmarks

+++ app/controllers/bookmarked_jobs_controller.rb
@@ -20,13 +20,13 @@ class BookmarkedJobsController < ApplicationController
   def create
     the_bookmarked_job = BookmarkedJob.new
     the_bookmarked_job.applicant_id = current_applicant.id
-    the_bookmarked_job.job_posting_id = params.fetch("query_job_posting_id")
+    the_bookmarked_job.job_posting_id = params.fetch("job_posting_id")
 
     if the_bookmarked_job.valid?
       the_bookmarked_job.save
-      redirect_to("/bookmarked_jobs", { :notice => "Bookmarked job created successfully." })
+      redirect_to("/job_postings", { :notice => "Bookmarked job created successfully." })
     else
-      redirect_to("/bookmarked_jobs", { :alert => the_bookmarked_job.errors.full_messages.to_sentence })
+      redirect_to("/job_postings", { :alert => the_bookmarked_job.errors.full_messages.to_sentence })
     end
   end
 
@@ -34,23 +34,32 @@ class BookmarkedJobsController < ApplicationController
     the_id = params.fetch("path_id")
     the_bookmarked_job = BookmarkedJob.where({ :id => the_id }).at(0)
 
-    the_bookmarked_job.applicant_id = params.fetch("query_applicant_id")
-    the_bookmarked_job.job_posting_id = params.fetch("query_job_posting_id")
+    if the_bookmarked_job.nil?
+      redirect_to("/job_postings", { :alert => "Bookmarked job not found." })
+      return
+    end
+
+    the_bookmarked_job.applicant_id = current_applicant.id
+    the_bookmarked_job.job_posting_id = params.fetch("job_posting_id")
 
     if the_bookmarked_job.valid?
       the_bookmarked_job.save
-      redirect_to("/bookmarked_jobs/#{the_bookmarked_job.id}", { :notice => "Bookmarked job updated successfully."} )
+      redirect_to("/job_postings", { :notice => "Bookmarked job updated successfully."} )
     else
-      redirect_to("/bookmarked_jobs/#{the_bookmarked_job.id}", { :alert => the_bookmarked_job.errors.full_messages.to_sentence })
+      redirect_to("/job_postings", { :alert => the_bookmarked_job.errors.full_messages.to_sentence })
     end
   end
 
   def destroy
-    the_id = params.fetch("path_id")
-    the_bookmarked_job = BookmarkedJob.where({ :id => the_id }).at(0)
+    the_job_posting_id = params.fetch("path_id")
+    the_bookmarked_job = current_applicant.bookmarked_jobs.where({ :job_posting_id => the_job_posting_id }).at(0)
+
+    if the_bookmarked_job.nil?
+      redirect_to("/job_postings", { :alert => "Bookmarked job not found." })
+      return
+    end
 
     the_bookmarked_job.destroy
-
-    redirect_to("/bookmarked_jobs", { :notice => "Bookmarked job deleted successfully."} )
+    redirect_to("/job_postings", { :notice => "Bookmarked job deleted successfully." })
   end
 end
